cambios DTo y reporte

+++ account-movement/src/main/java/com/microservicio/account/transaction/account_transaction/com/microservicio/account/transaction/account_transaction/services/ClientApiService.java
@@ -21,7 +21,7 @@ public class ClientApiService {
     }
 
     @SuppressWarnings("unchecked")
-    public List<Long> getAllIdClients() {
+    public List<Integer> getAllIdClients() {
         String url = CLIENT_PERSON_URL + "/clients/client/";
         return restTemplate.getForObject(url, List.class);
     }
